Update playwright tests
